moving test run to upload.js
